Update grafana-dmarc-reports.json
Update the titles of the over time graphs to match the Elasticsearch field names and pie chart names.

Those fields don't track if SPF or DKIM pass on their own. The domains used for the SPF/DKIM check that pass must **also** match/align with the message from header domain.